Use intptr_t

+++ src/post/goom/goomsl_heap.c
@@ -39,7 +39,7 @@ static void align_it(GoomHeap *_this, int alignment)
 {
   if ((alignment > 1) && (_this->number_of_arrays>0)) {
     void *last_array = _this->arrays[_this->number_of_arrays - 1];
-    long  last_address = (long)last_array + _this->consumed_in_last_array;
+    intptr_t  last_address = (intptr_t)last_array + _this->consumed_in_last_array;
     int   decal = (last_address % alignment);
     if (decal != 0) {
       _this->consumed_in_last_array += alignment - decal;
